Refactor plugin specific commands constants out
