+++ cypress/integration/plugins/index-management-dashboards-plugin/policies_spec.js
@@ -1,31 +1,9 @@
 /*
+ * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- *
- * Modifications Copyright OpenSearch Contributors. See
- * GitHub history for details.
  */
 
-/*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-import { BASE_PATH } from '../../../utils/constants';
-import { IM_PLUGIN_NAME } from "../../../utils/constants";
+import { BASE_PATH, IM_PLUGIN_NAME } from '../../../utils/constants';
 import samplePolicy from "../../../fixtures/plugins/index-management-dashboards-plugin/sample_policy";
 
 const POLICY_ID = "test_policy_id";
