+++ cypress/support/index.js
@@ -15,6 +15,8 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands';
+import '../utils/plugins/index-management-dashboards-plugin/commands'
+import '../utils/plugins/anomaly-detection-dashboards-plugin/commands'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
@@ -26,3 +28,16 @@ Cypress.on('uncaught:exception', (err) => {
     return false;
   }
 });
+
+// Proxy layer of AWS OpenSearch domain redirects to login page
+//  if you haven't authenticate
+if (Cypress.env('AWS_DOMAIN')) {
+  Cypress.Cookies.debug(false);
+  before(() => {
+    cy.login();
+  })
+
+  beforeEach(() => {
+    Cypress.Cookies.preserveOnce('security_authentication');
+  })
+}
